requirements.txt
mcp>=1.2.0
python-dotenv>=1.0.1
uvicorn>=0.30.0
fastapi>=0.115.0

pytest>=8.0.0
